+++ pypika/enums.py
@@ -125,9 +125,9 @@ class Dialects(Enum):
     ORACLE = "oracle"
     MSSQL = "mssql"
     MYSQL = "mysql"
-    POSTGRESQL = "postgressql"
+    POSTGRESQL = "postgresql"
     REDSHIFT = "redshift"
-    SQLLITE = "sqllite"
+    SQLLITE = "sqlite"
     SNOWFLAKE = "snowflake"